requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
